requirements.txt
pyYaml==3.12
paramiko==2.2.1
